fasgfsgddhfs

+++ src/main/java/ex28/App.java
@@ -11,19 +11,19 @@ public class App {
 
     public static void main(String[] args) {
         App myApp = new App();
-        int[] numArray = new int[5];
+        double[] numArray = new double[5];
 
         numArray = myApp.getNums(numArray);
-        int result = myApp.addNums(numArray);
+        double result = myApp.addNums(numArray);
         myApp.printOutput(result);
     }
 
-    public void printOutput(int output) {
-        System.out.printf("The total is %d.", output);
+    public void printOutput(double output) {
+        System.out.printf("The total is %.2f.", output);
     }
 
-    public int addNums(int[] arr) {
-        int res = 0;
+    public double addNums(double[] arr) {
+        double res = 0;
         for (int i = 0; i < 5; i++) {
             res += arr[i];
         }
@@ -31,7 +31,7 @@ public class App {
         return res;
     }
 
-    public int[] getNums(int[] arr) {
+    public double[] getNums(double[] arr) {
         for (int i = 0; i < 5; i++) {
             System.out.print("Enter a number: ");
             arr[i] = in.nextInt();
